pyproject.toml
[project]
name = "hikkinomore-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "google-genai>=1.29.0",
    "pydantic-settings>=2.10.1",
    "watchfiles>=1.1.0",
]
